+++ clikt/src/jvmTest/kotlin/com/github/ajalt/clikt/parameters/PromptOptionsTest.kt
@@ -1,20 +1,18 @@
 package com.github.ajalt.clikt.parameters
 
-import com.github.ajalt.clikt.core.BadParameterValue
-import com.github.ajalt.clikt.core.subcommands
-import com.github.ajalt.clikt.parameters.options.check
-import com.github.ajalt.clikt.parameters.options.nullableFlag
-import com.github.ajalt.clikt.parameters.options.option
-import com.github.ajalt.clikt.parameters.options.prompt
+import com.github.ajalt.clikt.core.*
+import com.github.ajalt.clikt.parameters.arguments.argument
+import com.github.ajalt.clikt.parameters.options.*
+import com.github.ajalt.clikt.parameters.transform.theme
 import com.github.ajalt.clikt.parameters.types.int
 import com.github.ajalt.clikt.testing.TestCommand
 import com.github.ajalt.clikt.testing.formattedMessage
 import com.github.ajalt.clikt.testing.parse
 import com.github.ajalt.clikt.testing.test
+import com.github.ajalt.mordant.terminal.ConversionResult
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.matchers.shouldBe
 import org.junit.Rule
-import org.junit.contrib.java.lang.system.SystemErrRule
 import org.junit.contrib.java.lang.system.SystemOutRule
 import org.junit.contrib.java.lang.system.TextFromStandardInputStream
 import kotlin.test.Test
@@ -30,15 +28,16 @@ class PromptOptionsTest {
 
     @Test
     fun `command prompt`() {
-        stdin.provideLines("bar")
+        stdin.provideLines("bar", "1")
 
         class C : TestCommand() {
             override fun run_() {
                 prompt("Foo") shouldBe "bar"
+                prompt("Baz") { ConversionResult.Valid(it.toInt()) } shouldBe 1
             }
         }
         C().parse("")
-        stdout.logWithNormalizedLineSeparator shouldBe "Foo: "
+        stdout.logWithNormalizedLineSeparator shouldBe "Foo: Baz: "
     }
 
     @Test